+++ src/main/java/fuzs/puzzleslib/config/option/BooleanOption.java
@@ -3,16 +3,14 @@ package fuzs.puzzleslib.config.option;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.fml.config.ModConfig;
 
-import java.util.function.BiFunction;
-
-public class BooleanOption extends ConfigOption<Boolean> {
+public class BooleanOption extends SimpleConfigOption<Boolean> {
 
     BooleanOption(ForgeConfigSpec.ConfigValue<Boolean> value, ModConfig.Type type, BooleanOptionBuilder builder) {
 
         super(value, type, builder);
     }
 
-    public static class BooleanOptionBuilder extends ConfigOptionBuilder<Boolean> {
+    public static class BooleanOptionBuilder extends SimpleConfigOptionBuilder<Boolean> {
 
         BooleanOptionBuilder(String name, Boolean defaultValue) {
 
@@ -20,15 +18,15 @@ public class BooleanOption extends ConfigOption<Boolean> {
         }
 
         @Override
-        BiFunction<ForgeConfigSpec.ConfigValue<Boolean>, ModConfig.Type, ConfigOption<Boolean>> getFactory() {
-
-            return (value, type) -> new BooleanOption(value, type, this);
+        SimpleConfigOption<Boolean> createOption(ForgeConfigSpec.ConfigValue<Boolean> value, ModConfig.Type type) {
+            
+            return new BooleanOption(value, type, this);
         }
 
         @Override
         ForgeConfigSpec.ConfigValue<Boolean> getConfigValue(ForgeConfigSpec.Builder builder) {
 
-            return builder.define(this.name, (boolean) this.defaultValue);
+            return builder.define(this.name, this.defaultValue);
         }
 
     }