+++ src/main/java/fuzs/puzzleslib/config/option/ConfigOption.java
@@ -1,17 +1,22 @@
 package fuzs.puzzleslib.config.option;
 
+import com.google.common.base.Splitter;
 import com.google.common.collect.Lists;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.fml.config.ModConfig;
 
 import java.util.List;
-import java.util.function.BiFunction;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-public abstract class ConfigOption<T> {
+public abstract class ConfigOption<T, S> {
 
-    private final ForgeConfigSpec.ConfigValue<T> value;
+    /**
+     * splitter for {@link #split}
+     */
+    private static final Splitter DOT_SPLITTER = Splitter.on(".");
+
+    private final ForgeConfigSpec.ConfigValue<S> value;
     private final ModConfig.Type type;
     private final List<String> path;
     private final String name;
@@ -21,7 +26,7 @@ public abstract class ConfigOption<T> {
     private final List<Consumer<T>> syncConsumers;
     private final List<Runnable> reloadListeners;
 
-    ConfigOption(ForgeConfigSpec.ConfigValue<T> value, ModConfig.Type type, ConfigOptionBuilder<T> builder) {
+    ConfigOption(ForgeConfigSpec.ConfigValue<S> value, ModConfig.Type type, ConfigOptionBuilder<T, S> builder) {
 
         this.value = value;
         this.type = type;
@@ -36,8 +41,10 @@ public abstract class ConfigOption<T> {
 
     public T get() {
 
-        return this.value.get();
+        return this.convertValue(this.value.get());
     }
+
+    protected abstract T convertValue(S value);
 
     public boolean isType(ModConfig.Type type) {
 
@@ -92,7 +99,17 @@ public abstract class ConfigOption<T> {
         this.reloadListeners.add(runOnReload);
     }
 
-    public static abstract class ConfigOptionBuilder<T> {
+    /**
+     * split helper copied from {@link ForgeConfigSpec}
+     * @param path path to split
+     * @return split path as list
+     */
+    protected static List<String> split(String path) {
+
+        return Lists.newArrayList(DOT_SPLITTER.split(path));
+    }
+
+    public static abstract class ConfigOptionBuilder<T, S> {
 
         final String name;
         final T defaultValue;
@@ -107,33 +124,33 @@ public abstract class ConfigOption<T> {
             this.defaultValue = defaultValue;
         }
 
-        public ConfigOptionBuilder<T> comment(String... comment) {
+        public ConfigOptionBuilder<T, S> comment(String... comment) {
 
             this.comment = comment;
             return this;
         }
 
-        public ConfigOptionBuilder<T> restart() {
+        public ConfigOptionBuilder<T, S> restart() {
 
             this.restart = true;
             return this;
         }
 
-        public ConfigOptionBuilder<T> sync(Consumer<T> syncToField) {
+        public ConfigOptionBuilder<T, S> sync(Consumer<T> syncToField) {
 
             this.syncConsumers.add(syncToField);
             return this;
         }
 
-        public ConfigOptionBuilder<T> listen(Runnable runOnReload) {
+        public ConfigOptionBuilder<T, S> listen(Runnable runOnReload) {
 
             this.reloadListeners.add(runOnReload);
             return this;
         }
 
-        abstract BiFunction<ForgeConfigSpec.ConfigValue<T>, ModConfig.Type, ConfigOption<T>> getFactory();
+        abstract ForgeConfigSpec.ConfigValue<S> getConfigValue(ForgeConfigSpec.Builder builder);
 
-        abstract ForgeConfigSpec.ConfigValue<T> getConfigValue(ForgeConfigSpec.Builder builder);
+        abstract ConfigOption<T, S> createOption(ForgeConfigSpec.ConfigValue<S> value, ModConfig.Type type);
 
     }