+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v1.0.13-1.16.5] - 2021-08-04
+### Added
+- Added FloatOption and FloatOptionBuilder
+- Added proper ListOption and ListOptionBuilder
+- GenericOption can now be restricted via predicate and list with allowed values
+
 ## [v1.0.12-1.16.5] - 2021-07-30
 ### Added
 - Added FuelManager for registering item burn times